chat_api/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat_api"
version = "0.1.0"
description = "A simple chat API"
authors = [
    { name = "Amogh Madan", email = "amoghmadaan@gmail.com" },
    { name = "Phaneendra Pallapothu", email = "phaneendra.pallapothu@gmail.com" },
]
readme = "README.md"
keywords = ["chat", "api", "python"]
license = { text = "MIT" }
requires-python = ">=3.11,<3.13"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application",
]
dependencies = [
    "asgiref==3.7.2",
    "attrs==23.2.0",
    "Django==4.2.10",
    "django-filter==23.5",
    "django-guardian==2.4.0",
    "djangorestframework==3.14.0",
    "djangorestframework-camel-case==1.4.2",
    "drf-spectacular==0.27.1",
    "inflection==0.5.1",
    "jsonschema==4.21.1",
    "jsonschema-specifications==2023.12.1",
    "psycopg==3.1.18",
    "psycopg-binary==3.1.18",
    "pytz==2024.1",
    "PyYAML==6.0.1",
    "referencing==0.33.0",
    "rpds-py==0.18.0",
    "sqlparse==0.4.4",
    "typing_extensions==4.9.0",
    "tzdata==2024.1",
    "uritemplate==4.1.1",
]

[project.optional-dependencies]
all = [
    "attrs==23.2.0",
    "cachetools==5.3.2",
    "chardet==5.2",
    "colorama==0.4.6",
    "distlib==0.3.7",
    "django-extensions==3.2.3",
    "filelock==3.13.1",
    "gunicorn==21.2.0",
    "packaging==23.2",
    "platformdirs==4.1",
    "pluggy==1.3",
    "pyproject-api==1.6.1",
    "tox==4.12.1",
    "virtualenv==20.25",
    "Werkzeug==3.0.1",
]
development = [
    "attrs==23.2.0",
    "cachetools==5.3.2",
    "chardet==5.2",
    "colorama==0.4.6",
    "distlib==0.3.7",
    "django-extensions==3.2.3",
    "filelock==3.13.1",
    "packaging==23.2",
    "platformdirs==4.1",
    "pluggy==1.3",
    "pyproject-api==1.6.1",
    "tox==4.12.1",
    "virtualenv==20.25",
    "Werkzeug==3.0.1",
]
production = ["gunicorn==21.2.0"]


[tool.isort]
profile = "black"

[tool.black]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88


[tool.bandit]
exclude_dirs = [
    "./.eggs",
    "./.git",
    "./.tox",
    "./build",
    "./coverage",
    "./dist",
    "./venv",
    "./__pycache__",
    "*tests*",
]


[tool.mypy]
namespace_packages = true
explicit_package_bases = true
ignore_missing_imports = true
exclude = """venv"""
